создан компонент movie-list
