+++ src/components/film-item/film-item.jsx
@@ -1,7 +1,7 @@
-import React, { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom';
-import { FilmItemProps } from './film-item-props';
-import { getFilmById } from '../../mocks/films';
+import React from 'react';
+import {Link} from 'react-router-dom';
+import {FilmItemProps} from './film-item-props';
+import {getFilmById} from '../../mocks/films';
 import VideoPlayer from '../player/video-player';
 
 const FilmItem = (props) => {
@@ -12,9 +12,9 @@ const FilmItem = (props) => {
     window.location.reload();
   }
   const {filmName, posterImage, videoLink, isActive} = filmByIdResult;
-
+  console.log(isActive)
   return (
-    <>
+    <article className="small-movie-card catalog__movies-card">
       <div className="small-movie-card__image">
         <Link to={to} >
           <VideoPlayer videoLink={videoLink} posterImage={posterImage} isActive={isActive} />
@@ -23,7 +23,8 @@ const FilmItem = (props) => {
       <h3 className="small-movie-card__title">
         <Link className="small-movie-card__link" to={to}>{filmName}</Link>
       </h3>
-    </>
+    </article>
+
   );
 };
 FilmItem.propTypes = FilmItemProps;
