+++ src/index.js
@@ -4,7 +4,7 @@ import App from './components/app/app';
 import {getFilmIds, getFilmById} from './mocks/films.js';
 import comments from './mocks/review';
 
-const filmIds = getFilmIds();
+// const filmIds = getFilmIds();
 const filmById = getFilmById;
 
-ReactDOM.render(<App filmIds = {filmIds} filmById = {filmById} comments = {comments}/>, document.querySelector(`#root`));
+ReactDOM.render(<App filmById = {filmById} comments = {comments}/>, document.querySelector(`#root`));
